+++ CHANGES.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.4.1 (2015-06/19)
+------------------
+
+- Fixed attribute loss in non-`dict` when conversion to `dict`.
+- Now automatically adding missing attributes to mapped objects.
+
 0.4 (2015-05-16)
 ----------------